currently commited no changes has to be made
Overview: The RandomUserInfoGenerator is a powerful tool designed to generate random user data that can be used for a variety of applications, including testing, development, and simulations. It creates realistic profiles with a wide range of personal information, ensuring a diverse set of random details such as names, addresses, emails, and more.